+++ CHANGELOG.md
@@ -6,11 +6,23 @@ All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
-[0.25.1]
+[0.26.1]
 ********
 Updated
 -------
 * Resolved issues for the reverse transform
+[0.26.0]
+********
+Updated
+-------
+* Disabled the configuredBand check in Configure command to enable Band 5 observation with real dish
+* This is a branch release.
+
+[0.25.1]
+********
+Fixed
+-----
+* Fixed issue with unresponsive flag update to resolve SKB-1074
 
 [0.25.0]
 ********